+++ nextgisweb/resource/nodepkg/effective-permissions/index.js
@@ -1,0 +1,4 @@
+/** @entrypoint */
+import { EffectivePermissions } from "./EffectivePermissions";
+
+export default EffectivePermissions;